fix linters issues
