+++ main.rb
@@ -1,5 +1,5 @@
 require_relative 'app'
-require_relative 'librarySistem'
+require_relative 'library_system'
 
 library_system = LibrarySystem.new
 library_system.start
